requirements.txt
pymongo
coreapi
django_redis
django
djangorestframework
shortuuid
kombu==4.5
librabbitmq
bson
